Actualizacion del codigo

+++ src/main/java/com/techcommunityperu/techcommunityperu/service/impl/InscripcionServiceImpl.java
@@ -1,17 +1,17 @@
 package com.techcommunityperu.techcommunityperu.service.impl;
 
 import com.techcommunityperu.techcommunityperu.dto.InscripcionDTO;
-import com.techcommunityperu.techcommunityperu.exception.InscripcionNotFoundException;
+import com.techcommunityperu.techcommunityperu.exception.ResourceNotFoundException;
 import com.techcommunityperu.techcommunityperu.mapper.InscripcionMapper;
 import com.techcommunityperu.techcommunityperu.model.entity.Inscripcion;
 import com.techcommunityperu.techcommunityperu.model.entity.Usuario;
 import com.techcommunityperu.techcommunityperu.repository.InscripcionRepository;
 import com.techcommunityperu.techcommunityperu.repository.UsuarioRepository;
 import com.techcommunityperu.techcommunityperu.service.InscripcionService;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import lombok.RequiredArgsConstructor;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
@@ -36,35 +36,30 @@ public class InscripcionServiceImpl implements InscripcionService {
     @Transactional
     @Override
     public void cancelarInscripcion(Integer eventoId, Integer usuarioId) {
-        Inscripcion inscripcion = inscriptionRepository.findByEventoAndUsuario(eventoId, usuarioId)
-                .orElseThrow(() -> new InscripcionNotFoundException("No se encontró inscripción para cancelar."));
+        Inscripcion inscripcion = inscriptionRepository.findByEventoAndUsuario(eventoId, usuarioId);
 
-        inscriptionRepository.deleteByEventoAndUsuario(eventoId, usuarioId);
+        if (inscripcion != null) {
+            inscriptionRepository.deleteByEventoAndUsuario(eventoId, usuarioId);
+        } else {
+            throw new ResourceNotFoundException("No se encontró la inscripción para cancelar.");
+        }
     }
 
     @Override
     public Inscripcion obtenerInscripcionPorEventoYUsuario(Integer eventoId, Integer usuarioId) {
-        return inscriptionRepository.findByEventoAndUsuario(eventoId, usuarioId)
-                .orElseThrow(() -> new InscripcionNotFoundException("No se encontró inscripción para el evento y usuario especificado."));
+        return inscriptionRepository.findByEventoAndUsuario(eventoId, usuarioId);
     }
-
 
     @Transactional
     @Override
     public void crearInscripcion(InscripcionDTO inscripcionDTO) {
-        // Validar si el usuario existe
-        Optional<Usuario> usuarioOpt = usuarioRepository.findById(inscripcionDTO.getUsuario());
-        if (usuarioOpt.isEmpty()) {
-            throw new InscripcionNotFoundException("Usuario no encontrado");
-        }
+        Usuario usuario = usuarioRepository.findById(inscripcionDTO.getUsuario())
+                .orElseThrow(() -> new ResourceNotFoundException("Usuario no encontrado"));
 
-        // Crear la entidad de Inscripción y mapear los campos desde el DTO
-        Inscripcion nuevaInscripcion = new Inscripcion();
-        nuevaInscripcion.setUsuario(usuarioOpt.get());
-        nuevaInscripcion.setMonto(inscripcionDTO.getMontoPago());
-        nuevaInscripcion.setTipoPago(inscripcionDTO.getTipoPago());
-        nuevaInscripcion.setInscripcionStatus(inscripcionDTO.getStatus());
+        Inscripcion nuevaInscripcion = inscripcionMapper.toEntity(inscripcionDTO);
+        nuevaInscripcion.setUsuario(usuario);
 
         inscriptionRepository.save(nuevaInscripcion);
     }
 }
+
